+++ library/stereotyped/src/main/kotlin/de/lise/fluxflow/stereotyped/job/ReflectedJobDefinition.kt
@@ -10,6 +10,7 @@ import java.time.Instant
 class ReflectedJobDefinition<TInstance>(
     override val kind: JobKind,
     override val parameters: List<ParameterDefinition<*>>,
+    override val metadata: Map<String, Any>,
     private val instance: TInstance,
     private val jobCaller: JobCaller<TInstance>
 ) : JobDefinition {
@@ -34,6 +35,7 @@ class ReflectedJobDefinition<TInstance>(
             status,
             workflow,
             parameterList,
+            metadata,
             instance,
             jobCaller
         )