+++ library/stereotyped/src/test/kotlin/de/lise/fluxflow/stereotyped/job/ReflectedJobTest.kt
@@ -22,6 +22,7 @@ class ReflectedJobTest {
             JobStatus.Scheduled,
             mock {},
             emptyList(),
+            emptyMap(),
             jobModelInstance,
             jobCaller
         )
@@ -48,6 +49,7 @@ class ReflectedJobTest {
             JobStatus.Scheduled,
             mock {},
             emptyList(),
+            emptyMap(),
             Any(),
             jobCaller
         )